requirements.txt
fastapi
sqlalchemy
mysql-connector-python
pymysql
python-dotenv
psycopg2-binary
